1. 將更新指令機制添加至每次重啟後 2. 抽離web server 3. 抽離slash commands設定邏輯
